Update labels text for map size x and y, #48

+++ src/Components/ColliderComponent.h
@@ -36,8 +36,8 @@ class ColliderComponent: public Component {
             #endif
             collider.x = static_cast<int>(transform->position.x);
             collider.y = static_cast<int>(transform->position.y);
-            collider.w = transform->width * transform->scale;
-            collider.h = transform->height * transform->scale;
+            collider.w = static_cast<int>(transform->width * transform->scale);
+            collider.h = static_cast<int>(transform->height * transform->scale);
             destinationRectangle.x = collider.x - Game::camera.x;
             destinationRectangle.y = collider.y - Game::camera.y;                
         }
